improved code and cleaned up
