+++ microservices-dashboard-server/pom.xml
@@ -185,12 +185,6 @@
             <version>2.0.12.RELEASE</version>
             <optional>true</optional>
         </dependency>
-        <dependency>
-            <groupId>net.sourceforge.htmlunit</groupId>
-            <artifactId>htmlunit</artifactId>
-            <version>2.23</version>
-            <scope>test</scope>
-        </dependency>
         <!-- TEST -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
