+++ microservices-dashboard-server/src/main/java/be/ordina/msdashboard/security/filter/AuthIndexFilter.java
@@ -16,7 +16,6 @@
 
 package be.ordina.msdashboard.security.filter;
 
-import be.ordina.msdashboard.security.strategy.SecurityProtocol;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
@@ -26,14 +25,11 @@ import org.springframework.stereotype.Component;
 @Component
 public class AuthIndexFilter extends AuthFilter {
 
-    @Value("${msdashboard.index.security}")
-    private String securityProtocol = SecurityProtocol.NONE.name();
-
-    public AuthIndexFilter(String securityProtocol) {
-        super(securityProtocol);
-    }
-
-
+	@Override
+	@Value("${msdashboard.health.security}")
+	protected void setSecurityProtocol(String protocol) {
+		super.setSecurityProtocol(protocol);
+	}
 }
