+++ microservices-dashboard-server/src/main/java/be/ordina/msdashboard/security/strategies/JWTAuthentication.java
@@ -24,76 +24,76 @@ import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import javax.servlet.http.HttpServletRequest;
 
 public class JWTAuthentication extends AbstractAuthenticationToken {
-    private static final long serialVersionUID = -4809832298438307309L;
-    private final HttpServletRequest storedRequest;
-    private final Authentication userAuthentication;
+	private static final long serialVersionUID = -4809832298438307309L;
+	private final HttpServletRequest storedRequest;
+	private final Authentication userAuthentication;
 
-    public JWTAuthentication(HttpServletRequest storedRequest, Authentication userAuthentication) {
-        super(userAuthentication.getAuthorities());
-        this.storedRequest = storedRequest;
-        this.userAuthentication = userAuthentication;
-    }
+	public JWTAuthentication(HttpServletRequest storedRequest, Authentication userAuthentication) {
+		super(userAuthentication.getAuthorities());
+		this.storedRequest = storedRequest;
+		this.userAuthentication = userAuthentication;
+	}
 
-    public Object getCredentials() {
-        return "";
-    }
+	public Object getCredentials() {
+		return "";
+	}
 
-    public Object getPrincipal() {
-        return this.userAuthentication.getPrincipal();
-    }
+	public Object getPrincipal() {
+		return this.userAuthentication.getPrincipal();
+	}
 
-    public HttpServletRequest getStoredRequest() {
-        return this.storedRequest;
-    }
+	public HttpServletRequest getStoredRequest() {
+		return this.storedRequest;
+	}
 
-    public Authentication getUserAuthentication() {
-        return this.userAuthentication;
-    }
+	public Authentication getUserAuthentication() {
+		return this.userAuthentication;
+	}
 
-    public boolean isAuthenticated() {
-        return this.userAuthentication.isAuthenticated();
-    }
+	public boolean isAuthenticated() {
+		return this.userAuthentication.isAuthenticated();
+	}
 
-    public void eraseCredentials() {
-        super.eraseCredentials();
-        if (this.userAuthentication != null && CredentialsContainer.class.isAssignableFrom(this.userAuthentication.getClass())) {
-            ((CredentialsContainer) CredentialsContainer.class.cast(this.userAuthentication)).eraseCredentials();
-        }
-    }
+	public void eraseCredentials() {
+		super.eraseCredentials();
+		if (this.userAuthentication != null && CredentialsContainer.class.isAssignableFrom(this.userAuthentication.getClass())) {
+			((CredentialsContainer) CredentialsContainer.class.cast(this.userAuthentication)).eraseCredentials();
+		}
+	}
 
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        } else if (!(o instanceof OAuth2Authentication)) {
-            return false;
-        } else if (!super.equals(o)) {
-            return false;
-        } else {
-            JWTAuthentication that = (JWTAuthentication) o;
-            if (!this.storedRequest.equals(that.storedRequest)) {
-                return false;
-            } else {
-                label33:
-                {
-                    if (this.userAuthentication != null) {
-                        if (this.userAuthentication.equals(that.userAuthentication)) {
-                            break label33;
-                        }
-                    } else if (that.userAuthentication == null) {
-                        break label33;
-                    }
+	public boolean equals(Object o) {
+		if (this == o) {
+			return true;
+		} else if (!(o instanceof OAuth2Authentication)) {
+			return false;
+		} else if (!super.equals(o)) {
+			return false;
+		} else {
+			JWTAuthentication that = (JWTAuthentication) o;
+			if (!this.storedRequest.equals(that.storedRequest)) {
+				return false;
+			} else {
+				label33:
+				{
+					if (this.userAuthentication != null) {
+						if (this.userAuthentication.equals(that.userAuthentication)) {
+							break label33;
+						}
+					} else if (that.userAuthentication == null) {
+						break label33;
+					}
 
-                    return false;
-                }
-                return true;
-            }
-        }
-    }
+					return false;
+				}
+				return true;
+			}
+		}
+	}
 
-    public int hashCode() {
-        int result = super.hashCode();
-        result = 31 * result + this.storedRequest.hashCode();
-        result = 31 * result + (this.userAuthentication != null ? this.userAuthentication.hashCode() : 0);
-        return result;
-    }
+	public int hashCode() {
+		int result = super.hashCode();
+		result = 31 * result + this.storedRequest.hashCode();
+		result = 31 * result + (this.userAuthentication != null ? this.userAuthentication.hashCode() : 0);
+		return result;
+	}
 }
