+++ microservices-dashboard-server/src/main/java/be/ordina/msdashboard/security/strategy/SecurityStrategy.java
@@ -18,7 +18,11 @@ package be.ordina.msdashboard.security.strategy;
 
 import org.springframework.stereotype.Component;
 
-import java.lang.annotation.*;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
  * Enables the Strategy pattern for adding security to your aggregator requests
@@ -26,12 +30,12 @@ import java.lang.annotation.*;
  * @author Kevin Van houtte
  */
 @Documented
-@Target({ ElementType.TYPE })
+@Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Component
 public @interface SecurityStrategy {
 
-    Class type();
+	Class type();
 
-    SecurityProtocol protocol() default SecurityProtocol.NONE;
+	SecurityProtocol protocol() default SecurityProtocol.NONE;
 }
