+++ microservices-dashboard-server/src/test/java/be/ordina/msdashboard/config/JWTAuthenticationInitializerInterceptor.java
@@ -15,10 +15,10 @@ import java.io.IOException;
  */
 public class JWTAuthenticationInitializerInterceptor implements ClientHttpRequestInterceptor {
 
-    @Override
-    public ClientHttpResponse intercept(HttpRequest httpRequest, byte[] bytes, ClientHttpRequestExecution clientHttpRequestExecution) throws IOException {
-        HttpHeaders headers = httpRequest.getHeaders();
-        headers.add("Authorization", "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ");
-        return clientHttpRequestExecution.execute(httpRequest, bytes);
-    }
+	@Override
+	public ClientHttpResponse intercept(HttpRequest httpRequest, byte[] bytes, ClientHttpRequestExecution clientHttpRequestExecution) throws IOException {
+		HttpHeaders headers = httpRequest.getHeaders();
+		headers.add("Authorization", "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ");
+		return clientHttpRequestExecution.execute(httpRequest, bytes);
+	}
 }
