+++ microservices-dashboard-server/src/test/java/be/ordina/msdashboard/security/strategy/SecurityStrategyFactoryTest.java
@@ -18,8 +18,6 @@
 package be.ordina.msdashboard.security.strategy;
 
 import be.ordina.msdashboard.security.strategies.StrategyFactory;
-import be.ordina.msdashboard.security.strategy.SecurityProtocol;
-import be.ordina.msdashboard.security.strategy.SecurityStrategy;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.InjectMocks;
@@ -35,67 +33,67 @@ import static org.mockito.Mockito.when;
 
 public class SecurityStrategyFactoryTest {
 
-    @Mock
-    private ApplicationContext applicationContext;
+	@Mock
+	private ApplicationContext applicationContext;
 
-    @InjectMocks
-    private final StrategyFactory factory = new StrategyFactory(applicationContext);
+	@InjectMocks
+	private final StrategyFactory factory = new StrategyFactory(applicationContext);
 
-    private Map<String, Object> annotatedBeans = new HashMap<>();
+	private Map<String, Object> annotatedBeans = new HashMap<>();
 
-    @Before
-    public void init() {
-        MockitoAnnotations.initMocks(this);
+	@Before
+	public void init() {
+		MockitoAnnotations.initMocks(this);
 
-        annotatedBeans.put("bean_oauth2", new OAuth2TestStrategy());
-        annotatedBeans.put("bean_jwt", new JWTTestStrategy());
-        annotatedBeans.put("bean_basic", new BasicTestStrategy());
-    }
+		annotatedBeans.put("bean_oauth2", new OAuth2TestStrategy());
+		annotatedBeans.put("bean_jwt", new JWTTestStrategy());
+		annotatedBeans.put("bean_basic", new BasicTestStrategy());
+	}
 
-    @Test
-    public void shouldGetOAuth2Strategy() {
-        when(applicationContext.getBeansWithAnnotation(SecurityStrategy.class)).thenReturn(annotatedBeans);
+	@Test
+	public void shouldGetOAuth2Strategy() {
+		when(applicationContext.getBeansWithAnnotation(SecurityStrategy.class)).thenReturn(annotatedBeans);
 
-        factory.init();
+		factory.init();
 
-        TestStrategy strategy = factory.getStrategy(TestStrategy.class, SecurityProtocol.OAUTH2);
-        assertEquals("SecurityStrategy returned was for the wrong profile", strategy.getClass(), OAuth2TestStrategy.class);
-    }
+		TestStrategy strategy = factory.getStrategy(TestStrategy.class, SecurityProtocol.OAUTH2);
+		assertEquals("SecurityStrategy returned was for the wrong profile", strategy.getClass(), OAuth2TestStrategy.class);
+	}
 
-    @Test
-    public void shouldGetJWTStrategy() {
-        when(applicationContext.getBeansWithAnnotation(SecurityStrategy.class)).thenReturn(annotatedBeans);
+	@Test
+	public void shouldGetJWTStrategy() {
+		when(applicationContext.getBeansWithAnnotation(SecurityStrategy.class)).thenReturn(annotatedBeans);
 
-        factory.init();
+		factory.init();
 
-        TestStrategy strategy = factory.getStrategy(TestStrategy.class, SecurityProtocol.JWT);
-        assertEquals("SecurityStrategy returned was for the wrong profile", strategy.getClass(), JWTTestStrategy.class);
-    }
+		TestStrategy strategy = factory.getStrategy(TestStrategy.class, SecurityProtocol.JWT);
+		assertEquals("SecurityStrategy returned was for the wrong profile", strategy.getClass(), JWTTestStrategy.class);
+	}
 
-    @Test
-    public void shouldGetBasicStrategy() {
-        when(applicationContext.getBeansWithAnnotation(SecurityStrategy.class)).thenReturn(annotatedBeans);
+	@Test
+	public void shouldGetBasicStrategy() {
+		when(applicationContext.getBeansWithAnnotation(SecurityStrategy.class)).thenReturn(annotatedBeans);
 
-        factory.init();
+		factory.init();
 
-        TestStrategy strategy = factory.getStrategy(TestStrategy.class, SecurityProtocol.BASIC);
-        assertEquals("SecurityStrategy returned was for the wrong profile", strategy.getClass(), BasicTestStrategy.class);
-    }
+		TestStrategy strategy = factory.getStrategy(TestStrategy.class, SecurityProtocol.BASIC);
+		assertEquals("SecurityStrategy returned was for the wrong profile", strategy.getClass(), BasicTestStrategy.class);
+	}
 
 
-    private interface TestStrategy {
-    }
+	private interface TestStrategy {
+	}
 
-    @SecurityStrategy(type = TestStrategy.class, protocol = SecurityProtocol.BASIC)
-    private static class BasicTestStrategy implements TestStrategy {
-    }
+	@SecurityStrategy(type = TestStrategy.class, protocol = SecurityProtocol.BASIC)
+	private static class BasicTestStrategy implements TestStrategy {
+	}
 
-    @SecurityStrategy(type = TestStrategy.class, protocol = SecurityProtocol.JWT)
-    private static class JWTTestStrategy implements TestStrategy {
-    }
+	@SecurityStrategy(type = TestStrategy.class, protocol = SecurityProtocol.JWT)
+	private static class JWTTestStrategy implements TestStrategy {
+	}
 
-    @SecurityStrategy(type = TestStrategy.class, protocol = SecurityProtocol.OAUTH2)
-    private static class OAuth2TestStrategy implements TestStrategy {
-    }
+	@SecurityStrategy(type = TestStrategy.class, protocol = SecurityProtocol.OAUTH2)
+	private static class OAuth2TestStrategy implements TestStrategy {
+	}
 
 }
