+++ microservices-dashboard-server/src/test/java/be/ordina/msdashboard/wiremock/InMemoryMockedConfiguration.java
@@ -22,29 +22,29 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class InMemoryMockedConfiguration {
 
-    @Bean
-    @ConditionalOnProperty(value = "pact-broker.url", havingValue = "https://localhost:8089")
-    protected InMemoryWireMock inMemoryWireMock() {
-        return new InMemoryWireMock();
-    }
+	@Bean
+	@ConditionalOnProperty(value = "pact-broker.url", havingValue = "https://localhost:8089")
+	protected InMemoryWireMock inMemoryWireMock() {
+		return new InMemoryWireMock();
+	}
 
-    @Bean
-    @ConditionalOnProperty(value = "msdashboard.health.security", havingValue = "basic")
-    protected InMemoryBasicWireMock inMemoryBasicWireMock() {
-        return new InMemoryBasicWireMock();
-    }
+	@Bean
+	@ConditionalOnProperty(value = "msdashboard.health.security", havingValue = "basic")
+	protected InMemoryBasicWireMock inMemoryBasicWireMock() {
+		return new InMemoryBasicWireMock();
+	}
 
-    @Bean
-    @ConditionalOnProperty(value = "msdashboard.health.security", havingValue = "jwt")
-    protected InMemoryJWTWireMock inMemoryJWTWireMock() {
-        return new InMemoryJWTWireMock();
-    }
+	@Bean
+	@ConditionalOnProperty(value = "msdashboard.health.security", havingValue = "jwt")
+	protected InMemoryJWTWireMock inMemoryJWTWireMock() {
+		return new InMemoryJWTWireMock();
+	}
 
-    @Bean
-    @ConditionalOnProperty(value = "msdashboard.health.security", havingValue = "oauth2")
-    protected InMemoryOAuth2WireMock inMemoryOAuth2WireMock() {
-        return new InMemoryOAuth2WireMock();
-    }
+	@Bean
+	@ConditionalOnProperty(value = "msdashboard.health.security", havingValue = "oauth2")
+	protected InMemoryOAuth2WireMock inMemoryOAuth2WireMock() {
+		return new InMemoryOAuth2WireMock();
+	}
 
 
 }
